1、修改动画：显示筛选类型界面动画

+++ sixLive/sixLive/Classes/Vendors/SIXPullRefresh/DLIntrospection/NSObject+DLIntrospection.m
@@ -170,7 +170,7 @@ static void getSuper(Class class, NSMutableString *result) {
         NSMutableArray *selParts = [[methodDescription componentsSeparatedByString:@":"] mutableCopy];
         NSInteger offset = 2; //1-st arg is object (@), 2-nd is SEL (:)
         
-        for (int idx = offset; idx < args; idx++) {
+        for (int idx = (int)offset; idx < args; idx++) {
             NSString *returnType = [NSString decodeType:method_copyArgumentType(methods[i], idx)];
             selParts[idx - offset] = [NSString stringWithFormat:@"%@:(%@)arg%d",
                                       selParts[idx - offset],
@@ -191,7 +191,7 @@ static void getSuper(Class class, NSMutableString *result) {
         [methodsDescription addObject:
          [NSString stringWithFormat:@"%@ (%@)%@",
           instance ? @"-" : @"+",
-#warning return correct type
+//#warning return correct type
           @"void",
           NSStringFromSelector(methods[i].name)]];
     }
